+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * `Pagination`
     * `Common`:
         * Extension `Pagination#reverse` has been added
+        * Factory `PaginationByIndexes`
+        * Shortcut `calculatePagesNumber` with reversed parameters
 
 ## 0.2.0